# api/docs/swagger.yaml
basePath: /v1
definitions:
  models.AllProject:
    properties:
      projects:
        items:
          $ref: '#/definitions/models.ProjectRes'
        type: array
    type: object
...
  models.FailureInfo:
    properties:
      description:
        type: string
      error: {}
      status_code:
        type: integer
    type: object
...
  models.ProjectReq:
    properties:
      developer_id:
        type: string
      file_url:
        type: string
      finished_date:
        example: 2006/01/02
        type: string
      project_name:
        type: string
      started_date:
        example: 2006/01/02
        type: string
    type: object
  models.ProjectRes:
    properties:
      developer_id:
        type: string
      file_url:
        type: string
      finished_date:
        example: 2006/01/02
        type: string
      id:
        type: integer
      project_name:
        type: string
      started_date:
        example: 2006/01/02
        type: string
      status:
        type: string
    type: object
  models.StatusUpdate:
    properties:
      id:
        type: integer
      status:
        type: string
    type: object
  models.Success:
    properties:
      message:
        type: string
      status_code:
        type: integer
    type: object
...
host: localhost:7777
info:
  contact:
    email: gofurovmurtazoxon@gmail.com
    name: Murtazoxon
    url: https://t.me/murtazokhon_gofurov
  description: This is Task api server
  termsOfService: 2 term OSG
  title: Online Service Group Task api
  version: "1.0"
paths:
...
  /prject:
    post:
      consumes:
      - application/json
      description: Through this api, can create project
      parameters:
      - description: CreateProject
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ProjectReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ProjectRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Create project
      tags:
      - Project
    put:
      consumes:
      - application/json
      description: Through this api, can update project
      parameters:
      - description: UpdateProject
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ProjectRes'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProjectRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Update project
      tags:
      - Project
  /prject/{id}:
    delete:
      consumes:
      - application/json
      description: Through this api, can delete a project
      parameters:
      - description: prject_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: DELETE a project
      tags:
      - Project
    get:
      consumes:
      - application/json
      description: Through this api, can get a project
      parameters:
      - description: prject_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProjectRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Get a project
      tags:
      - Project
  /prject/status:
    put:
      consumes:
      - application/json
      description: Through this api, can update project status
      parameters:
      - description: UpdateProjectStatus
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.StatusUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StatusUpdate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Update project status
      tags:
      - Project
  /prjects:
    get:
      consumes:
      - application/json
      description: Through this api, can get projects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AllProject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Get projects
      tags:
      - Project
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"